Custom JSON and RAW Generation for Day 3
* Updated all related data types for Answer's part2 from Int to Double for larger data sets

~ Day3Controller
   + GET.route("json", ":num")
   + GET.route("raw", ":num")
   ~ json(_:) updated to work with the new routes
   ~ raw(_:) updated to work with the new routes

+++ Sources/App/Extensions/Array.swift
@@ -9,11 +9,11 @@ import Foundation
 
 // MARK: Array<Int>
 extension Array where Element == Int {
-    func product() -> Int {
-        var product = self[0]
+    func product() -> Double {
+        var product = Double(self[0])
         
         for i in 1..<self.count {
-            product *= self[i]
+            product *= Double(self[i])
         }
         
         return product
